Crate air_test_utils

Modules

avm_runner

call_services

executed_state

key_utils

native_test_runner

ne_vec

prelude

raw_outcome

test_runner

wasm_test_runner

Macros

_trace_value_body

assert_next_pks

call_vm

checked_call_vm

failed

rc

scalar

scalar_tracked

stream

stream_tracked

unused

Please note that unused_tracked does not exist as unused is never tracked.

Structs

AVM

AVMConfig

Describes behaviour of the AVM.

AVMMemoryStats

Return statistic of AVM server Wasm module heap footprint.

AVMOutcome

AnomalyData

ApResult

Describes result of applying functor apply to streams.

CallRequestParams

Contains arguments of a call instruction and all other necessary information required for calling a service.

CallServiceFailed

CallServiceResult

Represents an executed host function result.

CanonCidAggregate

The type Canon trace CID refers to.

CanonResultCidAggregate

CidInfo

CidStore

Stores CID to Value corresponance.

CidTracker

ErrorAVMOutcome

ExecutionTrace

FoldResult

FoldSubTraceLore

Let’s consider an example of trace that could be produces by the following fold: (fold $stream v (call 1) (call 2) (next v) (call 3) (call 4) ) x Having started with stream with two elements {v1, v2} the resulted trace would looks like [(1) (2)] [(1) (2)] [(3) (4)] [(3) (4)] <— the sequence of call states v1 v2 v2 v1 <–– corresponding values from $stream that the iterable v had at the moment of call

GenerationIdx

InterpreterData

The AIR interpreter could be considered as a function f(prev_data: InterpreterData, current_data: InterpreterData, … ) -> (result_data: InterpreterData, …). This function receives prev and current data and produces a result data. All these data have the following format.

ParResult

ParticleParameters

Represents parameters obtained from a particle.

RawAVMOutcome

This struct is very similar to AVMOutcome, but keeps error_code and error_msg for test purposes.

SecurityTetraplet

Describes an origin that set corresponding value.

ServiceResultCidAggregate

A proof of service result execution result.

SubTraceDesc

Descriptor of a subtrace inside execution trace.

TracePos

Versions

Enums

AVMError

CallResult

CallSeDeErrors

CanonResult

Contains ids of element that were on a stream at the moment of an appropriate canon call.

ExecutedState

HostImportError

IType

Represents the types supported by WIT.

IValue

A WIT value.

Provenance

Sender

ValueRef

Constants

CALL_SERVICE_SUCCESS

Traits

DataStore

This trait is used for

Functions

check_error

data_from_result

data_version

into_raw_result

is_interpreter_succeded

print_trace

raw_data_from_trace

raw_data_from_trace_with_canon

trace_from_result

Type Aliases

AVMDataStore

AVMResult

CallRequests

CallResults

CallServiceClosure

FoldLore

This type represents all information in an execution trace about states executed during a fold execution.

HostExportedFunc

HostImportDescriptor

JValue

PosType

TraceLen